# feat: optimize batch processing and add comprehensive loading UX